apdate hw-2
